Separated profile tabs into components
