+++ frontend/src/components/Groups.jsx
@@ -1,0 +1,8 @@
+import { useState } from 'react';
+import styled from "styled-components";
+
+const Groups = () => (
+  <div>
+    <p>This is where the study groups will display</p>
+  </div>
+);
