+++ frontend/src/components/Schedule.jsx
@@ -1,0 +1,8 @@
+import { useState } from 'react';
+import styled from "styled-components";
+
+const Schedule = () => (
+  <div>
+    <p>This is where the class schedule will display</p>
+  </div>
+);
