+++ frontend/src/pages/Profile.jsx
@@ -2,6 +2,13 @@ import { useState, useEffect } from 'react';
 import { useParams } from 'react-router-dom';
 import styled from "styled-components";
 import ClassManager from "../components/ClassManager";
+import Friends from '../components/Friends';
+import Groups from '../components/Groups';
+import Schedule from '../components/Schedule';
+import Edit from '../components/Edit';
+
+
+
 
 const Container = styled.div`
   margin: auto;
@@ -9,7 +16,6 @@ const Container = styled.div`
   border: 3px solid var(--tertiary);
   border-radius: 1rem;
 `;
-
 
 const PageContainer = styled.body`
   display: flex;
@@ -105,213 +111,6 @@ const Content = styled.div`
   align-items: flex-start;
   justify-content: flex-start;
 `;
-
-const Friends = ({userId}) => {
-  const [email, setEmail] = useState('');
-  const [search, setSearch] = useState(null);
-  const [error, setError] = useState('');
-  const [success, setSuccess] = useState('');
-
-  const handleSearch = async() => {
-    setError('');
-    setSuccess('');
-    setSearch(null);
-
-    try {
-      const friends = await fetch(`http://localhost:3010/v0/users/searchFriend?userId=${userId}&email=${email}`, {
-        method: 'GET',
-        headers: {
-          'Authorization': `Bearer ${localStorage.getItem('accessToken')}`,
-          'Content-Type': 'application/json',
-        },
-      });
-
-      if (friends.status === 200) {
-        const fData = await friends.json();
-        if (fData) {
-          setError('User is already a friend');
-        } else {
-          setError('Server error');
-        }
-        return;
-      }
-
-      const response = await fetch(`http://localhost:3010/v0/users/search?email=${email}`, {
-        method: 'GET',
-        headers: {
-          'Authorization': `Bearer ${localStorage.getItem('accessToken')}`,
-          'Content-Type': 'application/json',
-        },
-      });
-
-      if (response.status === 200) {
-        const data = await response.json();
-        setSearch(data.user);
-        setError('');
-      } else if (response.status === 404) {
-        setError('User not found');
-        setSearch(null);
-      } else {
-        setError('Server error');
-        setSearch(null);
-      }
-    } catch (err) {
-      console.error(err);
-      setError('Failed to search for user');
-      setSearch(null);
-    }
-  };
-
-  const handleAdd = async () => {
-    if (!search) return
-
-    setError('');
-    setSuccess('');
-    try {
-      const response = await fetch(`http://localhost:3010/v0/users/addFriend`, {
-        method: 'PUT',
-        headers: {
-          'Authorization': `Bearer ${localStorage.getItem('accessToken')}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify({userId: userId, id: search.id, full_name: search.full_name, email: search.email})
-      });
-
-      if (response.status === 201) {
-        setSuccess(`${search.id} added!`);
-        setError('');
-        setSearch(null);
-      } else {
-        setError('Server error');
-        setSuccess('');
-      }
-    } catch (err) {
-      console.error(err);
-      setError('Failed to add');
-    }
-  };
-
-  return (
-    <div>
-    <Container>
-      <Label>Search by Email:</Label>
-      <input 
-        type="email"
-        placeholder="Enter email"
-        value={email}
-        onChange={(e) => setEmail(e.target.value)}
-      />
-      <button onClick={handleSearch}>Search</button>
-    </Container>
-    <br></br>
-    <div>
-      {search && (
-        <div>
-          <p>User found: {search.full_name}</p>
-          <button onClick={handleAdd}>Add Friend</button>
-        </div>
-      )}
-    </div>
-    {error && <p style={{color: 'red'}}>{error}</p>}
-    {success && <p style={{color: 'green'}}>{success}</p>}
-    </div>
-  );
-};
-
-const Groups = () => (
-  <div>
-    <p>This is where the study groups will display</p>
-  </div>
-);
-
-const Schedule = () => (
-  <div>
-    <p>This is where the class schedule will display</p>
-  </div>
-);
-
-const Edit = ({ profileData, setProfileData, setUpdateTrigger }) => {
-  const [formData, setFormData] = useState({id: profileData.id, full_name: profileData.full_name, bio: profileData.bio});
-  const [error, setError] = useState('');
-  const [success, setSuccess] = useState('');
-
-  const handleChange = (e) => {
-    const { id, value } = e.target;
-    setFormData( (prev) => ({...prev, [id]: value }));
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    setError('');
-    setSuccess('');
-
-    try {
-      const response = await fetch (`http://localhost:3010/v0/profile/${profileData.id}`, {
-        method: 'PUT',
-        headers: {
-          'Authorization': `Bearer ${localStorage.getItem('accessToken')}`,
-          'Content-Type': 'application/json',
-        },
-        body: JSON.stringify(formData),
-      });
-
-      if (response.status === 200) {
-        setUpdateTrigger((prev) => prev + 1);
-        setProfileData((prev) => ({...prev, full_name: formData.full_name, bio: formData.bio}));
-        setSuccess('Profile updated');
-        setError('');
-      } else if (response.status === 400) {
-        setError('Invalid data');
-        setSuccess('');
-      } else {
-        setError('Unexpected error');
-        setSuccess('');
-      }
-    } catch (err) {
-      console.error(err);
-      setError('Failed to update profile');
-      setSuccess('');
-    }
-  };
-
-  return (
-    <EditContainer>
-    <Container>
-    <h3>Edit Profile</h3>
-    <form onSubmit={handleSubmit}>
-      <div>
-        <Label>Username:</Label>
-      </div>
-      <div>
-        <input
-          id="full_name"
-          type="text"
-          value={formData.full_name}
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <Label>Bio:</Label>
-      </div>
-      <div>
-        <textarea
-          id="bio"
-          value={formData.bio}
-          onChange={handleChange}
-        />
-      </div>
-      <div>
-        <br></br>
-        <button type="submit">Save Changes</button>
-        {error && <p style={{ color: 'red' }}>{error}</p>}
-        {success && <p style={{ color: 'green' }}>{success}</p>}
-      </div>
-    </form>
-    </Container>
-    <ClassManager  userId={formData.id}/>
-    </EditContainer>
-  );
-};
 
 function Profile() {
   const { userId } = useParams();
